+++ src/main/java/com/ex/newWeb/Controller/SongController.java
@@ -6,6 +6,7 @@ import com.ex.newWeb.models.PlayList;
 import com.ex.newWeb.models.Song;
 import com.ex.newWeb.models.UserEntity;
 import com.ex.newWeb.security.SecurityUtil;
+import com.ex.newWeb.service.PlayListService;
 import com.ex.newWeb.service.SongService;
 import com.ex.newWeb.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,13 +24,22 @@ import java.util.List;
 public class SongController {
     private SongService songService;
     private UserService userService;
+    private PlayListService playListService;
 
     @Autowired
-    public SongController(SongService songService, UserService userService) {
+    public SongController(SongService songService, UserService userService,PlayListService playListService) {
         this.songService = songService;
         this.userService = userService;
+        this.playListService = playListService;
     }
 
+
+    @GetMapping("/allSongs")
+    public String listAllSongs(Model model){
+        List<SongDto> songs = songService.findAllSongs();
+        model.addAttribute("songs", songs);
+        return "all-songs-list";
+    }
 
     @GetMapping("/songs")
     public String songs(Model model){
@@ -38,26 +48,20 @@ public class SongController {
         return "songs-list";
     }
 
-    @GetMapping("/allSongs")
-    public String listAllSongs(Model model){
-
-        UserEntity user = new UserEntity();
-        List<SongDto> songs = songService.findAllSongs();
-        String username = SecurityUtil.getSessionUser();
-        if(username != null){
-            user = userService.findByUsername(username);
-            model.addAttribute("user",user);
-        }
-        model.addAttribute("user",user);
-        model.addAttribute("songs", songs);
-        return "all-songs-list";
-    }
     @GetMapping("/allSongs/search")
     public String searchAllSongs(@RequestParam(value = "query") String query, Model model){
         List<SongDto> songs = songService.searchAllSong(query);
         model.addAttribute("songs", songs);
         return "all-songs-list";
     }
+
+    @GetMapping("/songs/search")
+    public String searchSong(@RequestParam(value = "query") String query, Model model){
+        List<SongDto> songs = songService.searchSong(query);
+        model.addAttribute("songs", songs);
+        return "songs-list";
+    }
+
     @GetMapping("/songs/{songId}")
     public String viewSongs(@PathVariable("songId") Long songId, Model model){
         UserEntity user = new UserEntity();
@@ -72,28 +76,33 @@ public class SongController {
         return "song-detail";
     }
 
-    @GetMapping("/songs/search")
-    public String searchSong(@RequestParam(value = "query") String query, Model model){
-        List<SongDto> songs = songService.searchSong(query);
-        model.addAttribute("songs", songs);
-        return "songs-list";
-    }
     @GetMapping("/songs/{songId}/delete")
     public String songDelete(@PathVariable("songId") Long songId){
-        songService.deleteSong(songId);
+        String username = SecurityUtil.getSessionUser();
+        if(username != null) {
+            songService.deleteSong(songId);
+        }
         return "redirect:/songs";
     }
     @GetMapping("/songs/{songId}/copy")
     public String addCopySong(@PathVariable("songId") Long songId){
-        songService.saveCopySong(songId);
+        String username = SecurityUtil.getSessionUser();
+        if(username != null) {
+            songService.saveCopySong(songId);
+        }
         return "redirect:/songs";
     }
 
     @GetMapping("/songs/new")
     public String newSongForm(Model model){
-        Song song = new Song();
-        model.addAttribute("song", song);
-        return "song-create";
+        String username = SecurityUtil.getSessionUser();
+        if(username != null) {
+            Song song = new Song();
+            model.addAttribute("song", song);
+            return "song-create";
+        }else{
+            return "redirect:/home?error";
+        }
     }
 
 
@@ -109,9 +118,15 @@ public class SongController {
     }
     @GetMapping("/songs/{songId}/edit")
     public String editSong(@PathVariable("songId") Long songId, Model model){
-        SongDto songDto = songService.findSongById(songId);
-        model.addAttribute("song", songDto);
-        return "song-edit";
+        String username = SecurityUtil.getSessionUser();
+        if(username != null) {
+            SongDto songDto = songService.findSongById(songId);
+            if(username.equals(songDto.getCreatedBy().getUsername())) {
+                model.addAttribute("song", songDto);
+                return "song-edit";
+            }
+        }
+        return "redirect:/home?error";
     }
     @PostMapping("/songs/{songId}/edit")
     public String updateEvent(@PathVariable("songId") Long songId,
@@ -130,10 +145,18 @@ public class SongController {
 
     @GetMapping("/songs/{playListId}/addSong")
     public String addSongPlayList(@PathVariable("playListId") Long playListId, Model model){
-        List<SongDto> songs = songService.findYourSongs();
-        model.addAttribute("playListId", playListId);
-        model.addAttribute("songs", songs);
-        return "get-song";
+        String username = SecurityUtil.getSessionUser();
+        if(username != null) {
+            List<SongDto> songs = songService.findYourSongs();
+            if(songs != null) {
+                if (username.equals(songs.get(0).getCreatedBy().getUsername())) {
+                    model.addAttribute("playListId", playListId);
+                    model.addAttribute("songs", songs);
+                    return "get-song";
+                }
+            }
+        }
+        return "redirect:/home?error";
     }
     @GetMapping("/songs/{playListId}/addSong/search")
     public String searchSongInAdd(@RequestParam(value = "query") String query,
@@ -146,8 +169,14 @@ public class SongController {
     @GetMapping("/songs/{playListId}/addSong/{songId}")
     public String addSongToPlayList(@PathVariable("playListId") Long playListId,
                                     @PathVariable("songId") Long songId, Model model){
-        songService.addSongPlayList(songId,playListId);
-        return "redirect:/playLists/" + playListId;
+        String username = SecurityUtil.getSessionUser();
+        if(username != null) {
+            if (username.equals(songService.findSongById(songId).getCreatedBy().getUsername()) && username.equals(playListService.findPlayListById(playListId).getCreatedBy().getUsername())) {
+                songService.addSongPlayList(songId, playListId);
+                return "redirect:/playLists/" + playListId;
+            }
+        }
+        return "redirect:/home?error";
     }
 
     @GetMapping("/songs/{playListId}/remove/{songId}")
